move enum to DT package
